requirements.txt
aioredis==1.3.1
asgiref==3.3.1
channels==3.0.3
channels-redis==3.2.0
daphne==3.0.1
Django==3.1.5
lxml==4.6.2
pyserial==3.5
redis==3.5.3
svgelements==1.4.5
vpype @ git+https://github.com/abey79/vpype.git@a316c8aadeb8c2a4e014af634be4fb96cb959a76
